+++ tool.uart/src/main/java/nl/lxtreme/ols/tool/uart/impl/UARTDataSet.java
@@ -49,7 +49,7 @@ public final class UARTDataSet extends BaseDataSet<UARTData>
   // VARIABLES
 
   private int decodedSymbols;
-  private int bitLength;
+  private double bitLength;
   private int detectedErrors;
 
   private int baudRateExact;
@@ -88,7 +88,7 @@ public final class UARTDataSet extends BaseDataSet<UARTData>
    */
   public int getBaudRateExact()
   {
-    return this.getSampleRate() / this.bitLength;
+    return ( int )( this.getSampleRate() / this.bitLength );
   }
 
   /**
@@ -96,7 +96,7 @@ public final class UARTDataSet extends BaseDataSet<UARTData>
    * 
    * @return an average bit length, >= 0.
    */
-  public int getBitLength()
+  public double getBitLength()
   {
     return this.bitLength;
   }
@@ -202,7 +202,7 @@ public final class UARTDataSet extends BaseDataSet<UARTData>
    * @param aBitLength
    *          the bit length to set/add, should be >= 0.
    */
-  public void setSampledBitLength( final int aBitLength )
+  public void setSampledBitLength( final double aBitLength )
   {
     // If the given bit length is "much" smaller (smaller bit length means
     // higher baudrate) than the current one, switch to that one instead; this
@@ -217,16 +217,16 @@ public final class UARTDataSet extends BaseDataSet<UARTData>
       // Take the average as the current and the given bit lengths are "close"
       // to each other; ignore the given bit length otherwise, as it clobbers
       // our earlier results...
-      final int diff = Math.abs( aBitLength - this.bitLength );
+      final double diff = Math.abs( aBitLength - this.bitLength );
       if ( ( diff < 50 ) )
       {
-        this.bitLength = ( int )( ( aBitLength + this.bitLength ) / 2.0 );
+        this.bitLength = ( ( aBitLength + this.bitLength ) / 2.0 );
       }
       else
       {
         LOG.log( Level.INFO, "Ignoring sampled bit length ({0}) as it deviates "
             + "too much from current bit length ({1}).",
-            new Object[] { Integer.valueOf( aBitLength ), Integer.valueOf( this.bitLength ) } );
+            new Object[] { Double.valueOf( aBitLength ), Double.valueOf( this.bitLength ) } );
       }
     }
   }
